+++ VWOFmeSdk/Constants/Constants.cs
@@ -30,7 +30,7 @@ namespace VWOFmeSdk.Constants
 
         public const int SEED_VALUE = 1;
         public const int MAX_EVENTS_PER_REQUEST = 5000;
-        public const long DEFAULT_REQUEST_TIME_INTERVAL = 600; // 10 * 60 (secs) = 600 secs i.e. 10 minutes
+        public const int DEFAULT_REQUEST_TIME_INTERVAL = 600; // 10 * 60 (secs) = 600 secs i.e. 10 minutes
         public const int DEFAULT_EVENTS_PER_REQUEST = 100;
         public const string SDK_NAME = "vwo-fme-dotnet-sdk";
         public const int SETTINGS_EXPIRY = 10000000; // Changed from long to int