+++ CHANGELOG.md
@@ -5,11 +5,54 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2025-05-21
+
+### Added
+
+- Added support for `batchEventData` configuration to optimize network requests by batching multiple events together. This allows you to:
+
+  - Configure `requestTimeInterval` to flush events after a specified time interval
+  - Set `eventsPerRequest` to control maximum events per batch
+  - Implement `flushCallback` to handle batch processing results
+  - Manually trigger event flushing via `flushEvents()` method
+
+  - You can also manually flush events using the `flushEvents()` method:
+
+```c#
+   using VWOFmeSdk.Models;
+   using VWOFmeSdk.Interfaces.Batching;
+    IFlushInterface flushCallback = new FlushCallbackImpl();
+
+    var batchEventData = new BatchEventData
+    {
+        EventsPerRequest = 100,      // Send up to 100 events per
+        RequestTimeInterval = 60,   // Flush events every 60 seconds
+        FlushCallback = flushCallback,
+    };
+
+    var vwoInitOptions = new VWOInitOptions
+    {
+        SdkKey = "your_sdk_key",
+        AccountId = YOUR_ACCOUNT_ID,
+        BatchEventData = batchEventData
+    };
+
+    var vwoInstance = VWO.Init(vwoInitOptions);
+
+
+```
+
+- You can also manually flush events using the `flushEvents()` method:
+
+```c#
+  vwoInstance.flushEvents();
+```
+
 ## [1.7.0] - 2025-03-27
 
 ### Added
+
 - Added identifiable library to import uuid v5 methods instead of implementing in code
-- 
 
 ## [1.6.0] - 2025-03-12
 
@@ -26,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.0] - 2024-03-04
 
 ### Added
+
 - added new method `updateSettings` to update settings on the vwo client instance.
 
 ## [1.3.0] - 2024-11-11